ENH: Add functionalities and refactor graphics code.
 Now all graphics are in a separated subpackage.
