+++ pyqt-apps/siriushla/si_ap_sofb/graphics/__init__.py
@@ -1,0 +1,2 @@
+from .orbit import OrbitWidget
+from .correctors import CorrectorsWidget
